requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-admin==2.0.2
django-environ==0.11.2
psycopg2-binary==2.9.9

# desarrolllo
pytest==8.0.0
django-extensions==3.2.3
ipython==8.21.0
